git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/trunk@1421062 13f79535-47bb-0310-9956-ffa450edef68

+++ ambari-agent/src/main/python/ambari_agent/main.py
@@ -22,14 +22,15 @@ import logging
 import logging.handlers
 import code
 import signal
+from optparse import OptionParser
 import sys, traceback
 import os
 import time
 import ConfigParser
+import ProcessHelper
 from createDaemon import createDaemon
 from Controller import Controller
-from shell import getTempFiles
-from shell import killstaleprocesses 
+from shell import killstaleprocesses
 import AmbariConfig
 from security import CertificateManager
 from NetUtil import NetUtil
@@ -37,15 +38,10 @@ from NetUtil import NetUtil
 logger = logging.getLogger()
 agentPid = os.getpid()
 
-if 'AMBARI_PID_DIR' in os.environ:
-  pidfile = os.environ['AMBARI_PID_DIR'] + "/ambari-agent.pid"
-else:
-  pidfile = "/var/run/ambari/ambari-agent.pid"
-
 if 'AMBARI_LOG_DIR' in os.environ:
   logfile = os.environ['AMBARI_LOG_DIR'] + "/ambari-agent.log"
 else:
-  logfile = "/var/log/ambari/ambari-agent.log"
+  logfile = "/var/log/ambari-agent/ambari-agent.log"
 
 def signal_handler(signum, frame):
   #we want the handler to run only for the agent process and not
@@ -53,21 +49,7 @@ def signal_handler(signum, frame):
   if (os.getpid() != agentPid):
     os._exit(0)
   logger.info('signal received, exiting.')
-  try:
-    os.unlink(pidfile)
-  except Exception:
-    logger.warn("Unable to remove: "+pidfile)
-    traceback.print_exc()
-
-  tempFiles = getTempFiles()
-  for tempFile in tempFiles:
-    if os.path.exists(tempFile):
-      try:
-        os.unlink(tempFile)
-      except Exception:
-        traceback.print_exc()
-        logger.warn("Unable to remove: "+tempFile)
-  os._exit(0)
+  ProcessHelper.stopAgent()
 
 def debug(sig, frame):
     """Interrupt running process, and provide a python prompt for
@@ -85,6 +67,25 @@ def debug(sig, frame):
 
 def main():
   global config
+  parser = OptionParser()
+  parser.add_option("-v", "--verbose", dest="verbose", action="store_true", help="verbose log output", default=False)
+  (options, args) = parser.parse_args()
+
+  formatstr = "%(levelname)s %(asctime)s %(filename)s:%(lineno)d - %(message)s"
+  formatter = logging.Formatter(formatstr)
+  rotateLog = logging.handlers.RotatingFileHandler(logfile, "a", 10000000, 25)
+  rotateLog.setFormatter(formatter)
+  logger.addHandler(rotateLog)
+
+  if options.verbose:
+    logging.basicConfig(format=formatstr, level=logging.DEBUG, filename=logfile)
+    logger.setLevel(logging.DEBUG)
+  else:
+    logging.basicConfig(format=formatstr, level=logging.INFO, filename=logfile)
+    logger.setLevel(logging.INFO)
+
+  logger.debug("loglevel=logging.DEBUG")
+
   default_cfg = { 'agent' : { 'prefix' : '/home/ambari' } }
   config = ConfigParser.RawConfigParser(default_cfg)
   signal.signal(signal.SIGINT, signal_handler)
@@ -92,42 +93,41 @@ def main():
   signal.signal(signal.SIGUSR1, debug)
   if (len(sys.argv) >1) and sys.argv[1]=='stop':
     # stop existing Ambari agent
+    pid = -1
     try:
-      f = open(pidfile, 'r')
+      f = open(ProcessHelper.pidfile, 'r')
       pid = f.read()
       pid = int(pid)
       f.close()
       os.kill(pid, signal.SIGTERM)
       time.sleep(5)
-      if os.path.exists(pidfile):
+      if os.path.exists(ProcessHelper.pidfile):
         raise Exception("PID file still exists.")
       os._exit(0)
     except Exception, err:
-      os.kill(pid, signal.SIGKILL)
+      if pid == -1:
+        print ("Agent process is not running")
+      else:
+        os.kill(pid, signal.SIGKILL)
       os._exit(1)
 
   # Check if there is another instance running
-  if os.path.isfile(pidfile):
-    print("%s already exists, exiting" % pidfile)
+  if os.path.isfile(ProcessHelper.pidfile):
+    print("%s already exists, exiting" % ProcessHelper.pidfile)
     sys.exit(1)
   else:
     # Daemonize current instance of Ambari Agent
     #retCode = createDaemon()
     pid = str(os.getpid())
-    file(pidfile, 'w').write(pid)
-    
-  logger.setLevel(logging.INFO)
-  formatter = logging.Formatter("%(asctime)s %(filename)s:%(lineno)d - %(message)s")
-  rotateLog = logging.handlers.RotatingFileHandler(logfile, "a", 10000000, 10)
-  rotateLog.setFormatter(formatter)
-  logger.addHandler(rotateLog)
+    file(ProcessHelper.pidfile, 'w').write(pid)
+
   credential = None
 
   # Check for ambari configuration file.
   try:
     config = AmbariConfig.config
-    if os.path.exists('/etc/ambari/ambari.ini'):
-      config.read('/etc/ambari/ambari.ini')
+    if os.path.exists('/etc/ambari-agent/ambari-agent.ini'):
+      config.read('/etc/ambari-agent/ambari-agent.ini')
       AmbariConfig.setConfig(config)
     else:
       raise Exception("No config found, use default")
@@ -137,7 +137,8 @@ def main():
   killstaleprocesses()
 
   server_url = 'https://' + config.get('server', 'hostname') + ':' + config.get('server', 'url_port')
-  logger.info('Connecting to Server at: ' + server_url)
+  print("Connecting to the server at " + server_url + "...")
+  logger.info('Connecting to the server at: ' + server_url)
 
   # Wait until server is reachable
   netutil = NetUtil()
